separate validation request and update functions
