+++ app/Http/Requests/Store/UpdateStore.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Http\Requests\Store;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateStore extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'name' => 'required|max:25|unique:stores,name,' . $this->store->id,
+            'description' => 'required|string|max:500',
+        ];
+    }
+}
